app updates

+++ lib/api/src/modules/education_core/services/DeleteGradeService.ts
@@ -2,6 +2,7 @@ import { inject, injectable } from 'tsyringe';
 
 import AppError from '@shared/errors/AppError';
 
+import IClassroomsRepository from '@modules/schools/repositories/IClassroomsRepository';
 import IGradesRepository from '../repositories/IGradesRepository';
 
 type DeleteGradeRequest = {
@@ -12,6 +13,8 @@ type DeleteGradeRequest = {
 class DeleteGradeService {
   constructor(
     @inject('GradesRepository') private gradesRepository: IGradesRepository,
+    @inject('ClassroomsRepository')
+    private classroomsRepository: IClassroomsRepository,
   ) {}
 
   public async execute({ grade_id }: DeleteGradeRequest): Promise<void> {
@@ -19,6 +22,16 @@ class DeleteGradeService {
     if (!grade) {
       throw new AppError('Grade not found');
     }
+
+    const classrooms = await this.classroomsRepository.findAll({
+      grade_id,
+    });
+    if (classrooms.total > 0) {
+      throw new AppError(
+        'Exists classrooms linked with this grade. Cannot delete',
+      );
+    }
+
     await this.gradesRepository.delete(grade);
   }
 }
